end points/requirements.txt
azure-functions
azure-functions-worker
azure-core
azure-cosmos
